tools/gmail-ai-worker/requirements.txt
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.170.0,<3.0.0
